requirements.txt
flask
pandas
numpy
scikit-learn
torch
matplotlib
openai
python-dotenv
